requirements.txt
padaos
adapt-parser==0.3.0
padatious==0.4.5
fann2==1.0.7
python-dateutil
mtranslate
pronouncing
requests
psutil
bs4
all_the_chatbots>=0.1.1